Extract content into components
